requirements.txt
click
dlib
eos-py
filterpy
numpy
onnxruntime
opencv-python
oscpy
scipy
